requirements.txt
kivy
pyperclip
gcloud
google-cloud-bigquery
google-cloud-storage
google_auth_oauthlib
python-dotenv
google-cloud-secret-manager
cryptography
chardet